refactor: EventStoreHandler 제거, event 패키지 이동 #57
