+++ src/main/java/sogang/capstone/editking/domain/form/event/EventEntry.java
@@ -1,4 +1,4 @@
-package sogang.capstone.editking.domain.event;
+package sogang.capstone.editking.domain.form.event;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
