+++ src/main/java/sogang/capstone/editking/domain/form/event/EventForwarder.java
@@ -1,4 +1,4 @@
-package sogang.capstone.editking.domain.event;
+package sogang.capstone.editking.domain.form.event;
 
 import java.util.ArrayList;
 import java.util.List;
