+++ src/main/java/sogang/capstone/editking/infrastructure/event/EventEntryRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import sogang.capstone.editking.domain.event.EventEntry;
+import sogang.capstone.editking.domain.form.event.EventEntry;
 
 public interface EventEntryRepository extends JpaRepository<EventEntry, Long> {
