+++ src/main/java/sogang/capstone/editking/infrastructure/event/EventStoreImpl.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import sogang.capstone.editking.domain.event.EventEntry;
-import sogang.capstone.editking.domain.event.EventStore;
+import sogang.capstone.editking.domain.form.event.EventEntry;
+import sogang.capstone.editking.domain.form.event.EventStore;
 
 @Component
 @RequiredArgsConstructor
